Starting to add immunities

+++ src/engine/gameobject/units/poison/PoisonBuff.java
@@ -88,7 +88,7 @@ public class PoisonBuff extends Buff implements Poison{
     }
     
     public Buff clone(){
-        return new PoisonBuff(getDuration(), (int) getDamage());
+        return new PoisonBuff(getDuration(), getDamage());
     }
 
     @Override
